+++ Sources/Experiment/ExperimentAnalyticsProvider.swift
@@ -1,0 +1,15 @@
+//
+//  ExperimentAnalyticsProvider.swift
+//  Experiment
+//
+//  Created by Brian Giori on 7/23/21.
+//
+
+import Foundation
+
+/// Provides a analytics implementation for standard experiment events generated
+/// by the client (e.g. ``ExposureEvent``).
+@objc public protocol ExperimentAnalyticsProvider {
+    
+    @objc func track(_ event: ExperimentAnalyticsEvent)
+}